System Configuration enhancement.

+++ packages/Webkul/Core/src/Repositories/CoreConfigRepository.php
@@ -2,91 +2,92 @@
 
 namespace Webkul\Core\Repositories;
 
-use Webkul\Core\Eloquent\Repository;
+use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Storage;
-use Prettus\Repository\Traits\CacheableRepository;
+use Webkul\Core\Eloquent\Repository;
+use Webkul\Core\Contracts\CoreConfig;
 
 class CoreConfigRepository extends Repository
 {
-    use CacheableRepository;
-
     /**
-     * Specify Model class name
-     *
-     * @return mixed
+     * Specify model class name.
      */
-    function model()
+    public function model(): string
     {
-        return 'Webkul\Core\Contracts\CoreConfig';
+        return CoreConfig::class;
     }
 
     /**
-     * @param  array  $data
-     * @return \Webkul\Core\Contracts\CoreConfig
+     * Create core configuration.
+     *
      */
-    public function create(array $data)
+    public function create(array $data): void
     {
         unset($data['_token']);
 
         foreach ($data as $method => $fieldData) {
-            $recurssiveData = $this->recuressiveArray($fieldData , $method);
+            $recursiveData = $this->recursiveArray($fieldData, $method);
 
-            foreach ($recurssiveData as $fieldName => $value) {
-                $field = core()->getConfigField($fieldName);
-
-                if (getType($value) == 'array' && ! isset($value['delete'])) {
-                    $value = implode(",", $value);
+            foreach ($recursiveData as $fieldName => $value) {
+                if (
+                    gettype($value) == 'array'
+                    && ! isset($value['delete'])
+                ) {
+                    $value = implode(',', $value);
                 }
 
                 $coreConfigValue = $this->model
-                                    ->where('code', $fieldName)
-                                    ->get();
+                    ->where('code', $fieldName)
+                    ->get();
 
                 if (request()->hasFile($fieldName)) {
                     $value = request()->file($fieldName)->store('configuration');
                 }
 
                 if (! count($coreConfigValue)) {
-                    $this->model->create([
+                    parent::create([
                         'code'         => $fieldName,
                         'value'        => $value,
                     ]);
                 } else {
                     foreach ($coreConfigValue as $coreConfig) {
-                        Storage::delete($coreConfig['value']);
+                        if (request()->hasFile($fieldName)) {
+                            Storage::delete($coreConfig['value']);
+                        }
 
-                        if(isset($value['delete'])) {
-                            $this->model->destroy($coreConfig['id']);
+                        if (isset($value['delete'])) {
+                            parent::delete($coreConfig['id']);
                         } else {
-                            $coreConfig->update([
+                            parent::update([
                                 'code'         => $fieldName,
                                 'value'        => $value,
-                            ]);
+                            ], $coreConfig->id);
                         }
                     }
                 }
             }
         }
+
+        Event::dispatch('core.configuration.save.after');
     }
 
     /**
-     * @param  array  $formData
-     * @param  string  $method
-     * @return array
+     * Recursive array.
      */
-    public function recuressiveArray(array $formData, $method) {
+    public function recursiveArray(array $formData, string $method): array
+    {
         static $data = [];
 
-        static $recuressiveArrayData = [];
+        static $recursiveArrayData = [];
 
         foreach ($formData as $form => $formValue) {
-            $value = $method . '.' . $form;
+            $value = $method.'.'.$form;
 
             if (is_array($formValue)) {
                 $dim = $this->countDim($formValue);
 
                 if ($dim > 1) {
-                    $this->recuressiveArray($formValue, $value);
+                    $this->recursiveArray($formValue, $value);
                 } elseif ($dim == 1) {
                     $data[$value] = $formValue;
                 }
@@ -97,24 +98,21 @@ class CoreConfigRepository extends Repository
             $field = core()->getConfigField($key);
 
             if ($field) {
-                $recuressiveArrayData[$key] = $value;
+                $recursiveArrayData[$key] = $value;
             } else {
                 foreach ($value as $key1 => $val) {
-                    $recuressiveArrayData[$key . '.' . $key1] = $val;
+                    $recursiveArrayData[$key.'.'.$key1] = $val;
                 }
             }
         }
 
-        return $recuressiveArrayData;
+        return $recursiveArrayData;
     }
 
     /**
-     * Return dimension of array
-     *
-     * @param  array  $array
-     * @return int
-    */
-    public function countDim($array)
+     * Return dimension of the array.
+     */
+    public function countDim(array|string $array): int
     {
         if (is_array(reset($array))) {
             $return = $this->countDim(reset($array)) + 1;
